+++ docs/src/docs/components/chip-multi-select.mdx
@@ -62,8 +62,8 @@ export default Demo;`}
 ### Multiple Select Form
 
 <Demo code={`import { useId } from 'react';
-import { Button, FlexBox, ChipMultiSelect, Form, FormField, FormItem, FormControl, FormLabel, FormMessage } from '@wanteddev/wds';
-import { useForm } from 'react-hook-form';
+import { Button, FlexBox, ChipMultiSelect, FormField, FormControl, FormLabel, FormErrorMessage } from '@wanteddev/wds';
+import { useForm, Controller } from 'react-hook-form';
 
 const Demo = () => {
   const id = useId();
@@ -84,72 +84,71 @@ const Demo = () => {
     },
   ];
 
-  return (
-    <Form {...form}>
-      <FlexBox
-        as="form"
-        flexDirection="column"
-        gap="20px"
-        onSubmit={form.handleSubmit((v) => alert(JSON.stringify(v)))}
-      >
-        <FormField
-          control={form.control}
-          name="selected"
-          render={() => (
-            <FormItem>
-              <FormLabel>관심 분야</FormLabel>
-              <FlexBox gap="8px">
+  return ( 
+    <FlexBox
+      as="form"
+      flexDirection="column"
+      gap="20px"
+      onSubmit={form.handleSubmit((v) => alert(JSON.stringify(v)))}
+    >
+      <Controller
+        control={form.control}
+        name="selected"
+        render={({ formState }) => (
+          <FormField>
+            <FormLabel>관심 분야</FormLabel>
+            <FlexBox gap="8px">
 
-                {items.map((item) => (
-                  <FormField
-                    key={item.id}
-                    control={form.control}
-                    name="selected"
-                    rules={{
-                      required: {
-                        value: true,
-                        message: "필수 값입니다.",
-                      },
-                    }}
-                    render={({ field }) => {
-                      return (
-                        <FormItem
-                          key={item.id}
-                          flexDirection="row"
-                          alignItems="center"
-                        >
-                          <FormControl>
-                            <ChipMultiSelect
-                              {...field}
-                              checked={field.value.includes(item.id)}
-                              onCheckedChange={(checked) => {
-                                return checked
-                                  ? field.onChange([...field.value, item.id])
-                                  : field.onChange(
-                                      field.value.filter(
-                                        (value) => value !== item.id
-                                      )
-                                    );
-                              }}
-                            >
-                              {item.label}
-                            </ChipMultiSelect>
-                          </FormControl>
-                        </FormItem>
-                      );
-                    }}
-                  />
-                ))}
-              </FlexBox>
+              {items.map((item) => (
+                <Controller
+                  key={item.id}
+                  control={form.control}
+                  name="selected"
+                  rules={{
+                    required: {
+                      value: true,
+                      message: "필수 값입니다.",
+                    },
+                  }}
+                  render={({ field, formState }) => {
+                    return (
+                      <FormField
+                        key={item.id}
+                        flexDirection="row"
+                        alignItems="center"
+                      >
+                        <FormControl>
+                          <ChipMultiSelect
+                            {...field}
+                            checked={field.value.includes(item.id)}
+                            invalid={Boolean(formState.errors.selected)}
+                            onCheckedChange={(checked) => {
+                              return checked
+                                ? field.onChange([...field.value, item.id])
+                                : field.onChange(
+                                    field.value.filter(
+                                      (value) => value !== item.id
+                                    )
+                                  );
+                            }}
+                          >
+                            {item.label}
+                          </ChipMultiSelect>
+                        </FormControl>
+                      </FormField>
+                    );
+                  }}
+                />
+              ))}
+            </FlexBox>
 
-              <FormMessage />
-            </FormItem>
-          )}
-        />
+            <FormErrorMessage>{formState.errors.selected?.message}</FormErrorMessage>
+          </FormField>
+        )}
+      />
 
-        <Button type="submit">제출</Button>
-      </FlexBox>
-    </Form>
+      <Button type="submit">제출</Button>
+    </FlexBox>
   )
 }
