--- docs/src/docs/components/form-v3.mdx
@@ -1,53 +1,0 @@
----
-title: Form v3
-description: Form v3 Component
----
-
-# Form v3
-
-form 구조 및 접근성 처리를 다룹니다.
-각 컴포넌트별 활용법은 `TextField`, `TextArea`, `Checkbox`, `Radio` 등을 참고해주세요
-
-
-## 활용
-
-react-hook-form과 함께 활용하는 예시입니다.
-
-<Demo code={`import { Button, FlexBox, TextField, FormFieldV3, FormControlV3, FormLabelV3, FormMessageV3, FormErrorMessageV3 } from '@wanteddev/wds';
-import { Controller, useForm } from "react-hook-form";
-
-const Demo = () => {
-  const form = useForm();
-
-  return (
-    <FlexBox as="form" flexDirection="column" gap="12px" onSubmit={form.handleSubmit(v => alert(JSON.stringify(v)))}>
-      <Controller
-        control={form.control}
-        name="content"
-        rules={{
-          required: {
-            value: true,
-            message: '필수 값입니다.'
-          }
-        }}
-        render={({ field, formState }) => (
-          <FormFieldV3>
-            <FormLabelV3>
-              라벨
-            </FormLabelV3>
-            
-            <FormControlV3 error={formState.errors.content}>
-              <TextField {...field} width="300px" placeholder="값을 입력하세요." />
-            </FormControlV3>
-
-            <FormErrorMessageV3>{formState.errors.content?.message}</FormErrorMessageV3>
-          </FormFieldV3>
-        )}
-      />
-
-      <Button type="submit">제출</Button>
-    </FlexBox>
-  )
-}
-
-export default Demo;`}/>
