+++ packages/wds/package.json
@@ -70,23 +70,18 @@
     "object-path": "^0.11.8",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-hook-form": "^7.51.0",
     "react-remove-scroll": "^2.5.7",
     "zustand": "^4.5.2"
   },
   "peerDependencies": {
     "react": ">= 18",
-    "react-dom": ">= 18",
-    "react-hook-form": ">= 7"
+    "react-dom": ">= 18"
   },
   "peerDependenciesMeta": {
     "@types/react": {
       "optional": true
     },
     "@types/react-dom": {
-      "optional": true
-    },
-    "react-hook-form": {
       "optional": true
     }
   }
