--- packages/wds/src/components/form-v3/constants.ts
@@ -1,5 +1,0 @@
-export const FORM_FIELD_NAME = 'FormField-v3';
-export const FORM_LABEL_NAME = 'FormLabel-v3';
-export const FORM_CONTROL_NAME = 'FormControl-v3';
-export const FORM_MESSAGE_NAME = 'FormMessage-v3';
-export const FORM_ERROR_MESSAGE_NAME = 'FormErrorMessage-v3';
