+++ packages/wds/src/components/form/constants.ts
@@ -1,7 +1,5 @@
 export const FORM_FIELD_NAME = 'FormField';
-export const FORM_ITEM_NAME = 'FormItem';
 export const FORM_LABEL_NAME = 'FormLabel';
 export const FORM_CONTROL_NAME = 'FormControl';
-export const FORM_ERROR_MESSAGE_NAME = 'FormErrorMessage';
-export const FORM_DESCRIPTION_NAME = 'FormDescription';
 export const FORM_MESSAGE_NAME = 'FormMessage';
+export const FORM_ERROR_MESSAGE_NAME = 'FormErrorMessage';
